commit 20.12.2022 20:05

+++ app/src/main/java/com/example/nettoyeurs/WebServiceCreer.kt
@@ -7,7 +7,7 @@ import org.w3c.dom.NodeList
 import java.net.URL
 import javax.xml.parsers.DocumentBuilderFactory
 
-class WebServiceCreer(val session: Int,val signature: Int,val lon: Double, val lat: Double) {
+class WebServiceCreer(val session: Int,val signature: Long,val lon: Double, val lat: Double) {
     val TAG = "WSCreer"
 
     fun call(): ArrayList<Node>? {
